Refactor fixtures to own file
